Refactor: 사이드메뉴뷰 세팅뷰로 분리
